pyproject.toml
[tool.poetry]
name = "aipac"
version = "0.1.0"
description = ""
authors = ["saadi yousaf"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-bigquery = {extras = ["pandas"], version = "^3.27.0"}
pydantic = "^2.9.2"
google-cloud = "^0.34.0"
google-api-core = "^2.23.0"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
